# Improve performance by skipping hidden dirs