--- a/web/server/watch/imperative_shell.go
+++ b/web/server/watch/imperative_shell.go
@@ -5,6 +5,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 )
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -30,6 +31,10 @@ func YieldFileSystemItems(root string, excludedDirs []string) chan *FileSystemIt
 	go func() {
 		filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
 			if err != nil {
+				return filepath.SkipDir
+			}
+
+			if info.IsDir() && strings.HasPrefix(info.Name(), ".") {
 				return filepath.SkipDir
 			}