@riannucci 2016-09-28
This LGTM. I don't expect that folks are using .-prefixed subdirectories to house things they actually want goconvey to test. As I press the merge button, however, I just know the internet is going to tell me otherwise :).